change formatAmount to Long on IncomeListAdapter

+++ app/src/main/java/com/example/denip/nasyiatulaisyiyahfinance/income/IncomeListAdapter.kt
@@ -62,7 +62,7 @@ class IncomeListAdapter(context: Context?, incomes: ArrayList<IncomeModel>) : Re
         })
     }
 
-    private fun formatAmount(amount: Int): String =
+    private fun formatAmount(amount: Long): String =
         NumberFormat.getNumberInstance(Locale.US).format(amount)
 
     override fun getItemCount(): Int = incomes.size
